Align side image
The basic files that will be used in this project are created. A custom
font is also added, which will be used for the banner on top of the
image. The image that belongs the left side of the screen is aligned
appropriately using css, and credit is given on the bottom of the
image's div.