Make form wrap
The banner will move to the top of the screen when the screen
resolution is too small. This makes the website look better across
mobile devices. Lots of code is commented out in order to create this
functionality. Some code adjustments may need to be made, especially
for company logo on banner.