Add error messages to the appropriate fields
A red message is displayed under each required input, along with the
phone number input. These errors are supposed to be displayed on the
screen when the inputs are given invalid values, and disappear when
the inputted value validates. Some javascript is also changed, making
it so the text boxes aren't given a red border if the user enters
something and then removes it, leaving the input completely empty.